# Add caching for methods with coerced args