Add caching speeds up a bit some cases
Before
```
Benchmark                                  Mode  Cnt      Score     Error   Units
DatafakerExpressions.bothifyExpression    thrpt   10   2169.941 ± 45.504  ops/ms
DatafakerExpressions.letterifyExpression  thrpt   10   2104.585 ± 81.704  ops/ms
DatafakerExpressions.regexifyExpression   thrpt   10   1607.923 ± 34.328  ops/ms
```
After
```
Benchmark                                  Mode  Cnt      Score     Error   Units
DatafakerExpressions.bothifyExpression    thrpt   10   2283.968 ±  33.019  ops/ms
DatafakerExpressions.letterifyExpression  thrpt   10   2367.088 ±  61.098  ops/ms
DatafakerExpressions.regexifyExpression   thrpt   10   2140.930 ±  61.182  ops/ms
```